airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark==3.3.2
requests
apache-airflow-providers-google
scikit-learn
pandas
google-cloud-bigquery
google-cloud-storage
joblib
pyarrow